Implement libsql scorer api
